+++ lib/atome/renderers/browser/identity.rb
@@ -2,14 +2,14 @@
 
 generator = Genesis.generator
 
-generator.build_render(:browser_id) do |params|
-  if @browser_type == :style
-    prev_content = @browser_object.inner_html
-    new_content = prev_content.sub(@browser_object.id, params)
-    @browser_object.inner_html = new_content
-  end
-  browser_object.id = params if @atome[:id] != params
-end
+# generator.build_render(:browser_id) do |params|
+#   if @browser_type == :style
+#     prev_content = @browser_object.inner_html
+#     new_content = prev_content.sub(@browser_object.id, params)
+#     @browser_object.inner_html = new_content
+#   end
+#   browser_object.id = params if @atome[:id] != params
+# end
 
 generator.build_render(:browser_type) do |params|
   send("browser_#{params}", user_proc)
@@ -26,7 +26,8 @@ end
 
 generator.build_render(:browser_attached) do |children_found|
   children_found.each do |child_found|
-    # puts ">>>>> modeled:  browser_type #{@browser_type} parents_found #{children_found} , @browser_object, #{@browser_object}, self : #{self.id}"
+    # puts ">>>>> modeled:  browser_type #{@browser_type} parents_fo
+    # und #{children_found} , @browser_object, #{@browser_object}, self : #{self.id}"
     children_browser_type = grab(child_found).instance_variable_get('@browser_type')
     BrowserHelper.send("browser_attached_#{children_browser_type}", children_found, @browser_object, @atome)
   end