+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 
 ## [0.5.4.1.2] - 2023-05-20
 - added changelog for rubygems
+
+## [0.5.4.1.3] - 2023-05-21
+-version updated to  '0.5.4.1.3'
+vector now remove background-color if it exist
+vector now use class instead of a local style